Moved nested mount points tests to ViewFileSystemBaseTest
